Add files via upload
This is the music app where I built a webpage interface like a music app where user can play music pause and move to next and go to previous . There is also toggle of light/dark mode implemented. In this webpage music app , the tech stacks have been used are HTML, CSS AND JAVASCRIPT. I am looking forward to make it more responsive and user friendly and want to add extra features and functionality through REACTJS , NODEJS , MONGODB AND EXPRESSJS. This is the part of my internship at CODE ALPHA in web development.